+++ bin/test_client.rb
@@ -3,9 +3,16 @@ require 'notifications/client'
 require 'notifications/client/notification'
 require 'notifications/client/response_notification'
 require 'notifications/client/notification'
+require 'notifications/client/response_template'
+require 'notifications/client/template_collection'
 
 def main
   client = Notifications::Client.new(ENV['API_KEY'], ENV['NOTIFY_API_URL'])
+  test_get_template_by_id(client, ENV['EMAIL_TEMPLATE_ID'])
+  test_get_template_version(client, ENV['SMS_TEMPLATE_ID'], 1)
+  test_get_all_templates(client)
+  test_get_all_templates_filter_by_type(client)
+  test_generate_template_preview(client, ENV['EMAIL_TEMPLATE_ID'])
   email_notification = test_send_email_endpoint(client)
   sms_notification = test_send_sms_endpoint(client)
   test_get_notification_by_id_endpoint(client, email_notification.id, 'email')
@@ -13,6 +20,73 @@ def main
   test_get_all_notifications(client)
   p 'ruby client integration tests pass'
   exit 0
+end
+
+def test_get_template_by_id(client, id)
+  response = client.get_template_by_id(id)
+  test_template_response(response, 'test_get_template_by_id')
+end
+
+def test_get_template_version(client, id, version)
+  response = client.get_template_version(id, version)
+  test_template_response(response, 'test_get_template_version')
+end
+
+def test_get_all_templates(client)
+  response = client.get_all_templates()
+  unless response.is_a?(Notifications::Client::TemplateCollection) then
+    p 'failed test_get_all_templates response is not a Notifications::Client::TemplateCollection'
+    exit 1
+  end
+  unless response.collection.length >= 2 then
+    p 'failed test_get_all_templates, expected at least 2 templates returned.'
+    exit 1
+  end
+  test_template_response(response.collection[0], 'test_get_all_templates')
+  test_template_response(response.collection[1], 'test_get_all_templates')
+end
+
+def test_get_all_templates_filter_by_type(client)
+  response = client.get_all_templates({'type' => 'sms'})
+  unless response.is_a?(Notifications::Client::TemplateCollection) then
+    p 'failed test_get_all_templates response is not a Notifications::Client::TemplateCollection'
+    exit 1
+  end
+  unless response.collection.length >= 1 then
+    p 'failed test_get_all_templates, expected at least 2 templates returned.'
+    exit 1
+  end
+  test_template_response(response.collection[0], 'test_get_all_templates')
+end
+
+def test_generate_template_preview(client, id)
+
+  response = client.generate_template_preview(id, personalisation:Hash["name", "some name"])
+  test_template_preview(response)
+end
+
+def test_template_response(response, test_method)
+  unless response.is_a?(Notifications::Client::Template) then
+    p 'failed test_get_template_by_id response is not a Notifications::Client::Template'
+    exit 1
+  end
+  unless response.id.is_a?(String) then
+    p 'failed template id is not a String'
+    exit 1
+  end
+  field_should_not_be_nil(expected_fields_in_template_response, response, test_method)
+end
+
+def test_template_preview(response)
+  unless response.is_a?(Notifications::Client::TemplatePreview) then
+    p 'failed test_generate_template_preview response is not a Notifications::Client::TemplatePreview'
+    exit 1
+  end
+  unless response.id.is_a?(String) then
+    p 'failed template id is not a String'
+    exit 1
+  end
+  field_should_not_be_nil(expected_fields_in_template_preview, response, 'generate_template_preview')
 end
 
 def test_send_email_endpoint(client)
@@ -99,6 +173,23 @@ def field_should_be_nil(fields, obj, method_name)
   end
 end
 
+def expected_fields_in_template_response
+  %w(id
+     type
+     created_at
+     created_by
+     body
+     version
+)
+end
+
+def expected_fields_in_template_preview
+  %w(id
+     body
+     version
+     type
+)
+end
 
 def expected_fields_in_notification_response
   %w(id